Simplify member structure
Instead of having three lists, introduced two parameters on a member to indicate the state of this member.
Voter / Observer is now a type.
Connected is now the state.
These are used to filter the users by.
This reduces a lot of copying and setting/unsetting members on different lists.