Optimize data-flow
Combined all vote-related data into the "vote" response.
Some values where in the "story" object, which resulted in two events having to be sent for many triggers.
It would also reset the story name when a vote was changed or user left/entered the room. When changing the story name, this was very unwanted.